projects/beaglebone-gateway/requirements.txt
# Industrial Gateway Python Dependencies

# Core
pyyaml>=6.0
quart>=0.19.0
hypercorn>=0.15.0
aiohttp>=3.9.0

# Modbus TCP/RTU
pymodbus>=3.6.0

# Allen-Bradley EtherNet/IP
pycomm3>=1.2.0

# Siemens S7 (also requires libsnap7 compiled)
python-snap7>=1.3

# Mitsubishi MELSEC
pymcprotocol>=0.4.0

# OPC UA Server
asyncua>=1.0.0

# DF1 Protocol (Allen-Bradley serial)
# df1py3>=0.1.0  # Optional, install if needed

# Development/Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
